Added Particle.function cloudMode
Cloud functions are fairly core to the Particle system so I added a Particle Console function to change modes.  I haven't tested this code and I'm sure it could be cleaned up quite a bit since it duplicates the button function but I just wanted to get this in before Christmas!